Restructure flask app
