+++ cbt_logger/templates/Emotions.html
