+++ cbt_logger/templates/Event.html
