+++ cbt_logger/templates/Thoughts.html
