+++ cbt_logger/templates/emotions.js
