+++ cbt_logger/templates/index.html
