+++ cbt_logger/templates/register.html
