+++ run.py
@@ -1,0 +1,4 @@
+from cbt_logger import app
+
+if __name__ == "__main__":
+    app.run(port=8992, debug=True)
